INS-2272: fixes after moving files
