+++ ledger/proc/proc.go
@@ -28,7 +28,7 @@ type Dependencies struct {
 	WaitHot    func(*WaitHot) *WaitHot
 	GetIndex   func(*GetIndex) *GetIndex
 	SendObject func(p *SendObject) *SendObject
-	GetCodeRec func(*GetCodeRec) *GetCodeRec
+	GetCode    func(*GetCode) *GetCode
 }
 
 type ReturnReply struct {
